feat: Change node to use hub, Create node specific hub
