+++ packages/hub/src/index.ts
@@ -1,2 +1,3 @@
+export { Carrier, Layer } from './interfaces';
 export { Scope } from './scope';
 export { hubFromCarrier, Hub } from './hub';
